[Stdlib] Unify Int with SIMD
This both changes the definition of `Int` over to an alias for
`SIMD[DType.int, 1]` and
migrates some usages of `Int` to a new `SIMDSize` type.

This new type is required for parameter inference to work in cases where
`SIMD` arguments
are passed with inferred width. It should otherwise not be commonly
used. SIMDSize is essentially the same as the "old" Int type, but may be
backed by a `pop.scalar<index>` in the future.

MODULAR_ORIG_COMMIT_REV_ID: cb60548c595da4ad798e9d17f50d7e516f90fc1a

+++ KGEN/unittests/mojo-debug/inputs/stack_trace.mojo
@@ -17,7 +17,7 @@ struct Foo[X: TrivialRegisterPassable, Y: TrivialRegisterPassable]:
         return nested_function(val)
 
     def getFloat(self, x: Float32, y: Int) -> Float32:
-        var tmp = self.getParametrized[Float32](Float32(4.125 + x + y))
+        var tmp = self.getParametrized[Float32](Float32(4.125 + x + Float32(y)))
         stop_tail_call()
         return tmp
